requirements.txt
# Auto-generated from pyproject.toml — keep in sync
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
bleach==6.2.0
jinja2==3.1.4
python-multipart==0.0.22
pydantic[email]
